Refactoring in tests
- Create packages for ListView, SlideView tests and TestUtils
- Move Matchers to separated file
