--- src/app/src/androidTest/java/com/przyjaznyplanDisplayer/ApplicationTest.java
@@ -1,20 +1,0 @@
-/*
- * Copyright (c) 2016. Wydział Elektroniki, Telekomunikacji i Informatyki, Politechnika Gdańska
- * This program is free software; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation; either version 3 of the License, or   (at your option) any later version.
- *
- * Copy of GNU General Public License is available at http://www.gnu.org/licenses/gpl-3.0.html
- */
-
-package com.przyjaznyplanDisplayer;
-
-import android.app.Application;
-import android.test.ApplicationTestCase;
-
-/**
- * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
- */
-public class ApplicationTest extends ApplicationTestCase<Application> {
-    public ApplicationTest() {
-        super(Application.class);
-    }
-}
