--- src/app/src/androidTest/java/com/przyjaznyplanDisplayer/FirstTest.java
@@ -1,21 +1,0 @@
-/*
- * Copyright (c) 2016. Wydział Elektroniki, Telekomunikacji i Informatyki, Politechnika Gdańska
- * This program is free software; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation; either version 3 of the License, or   (at your option) any later version.
- *
- * Copy of GNU General Public License is available at http://www.gnu.org/licenses/gpl-3.0.html
- */
-
-package com.przyjaznyplanDisplayer;
-
-import android.test.ActivityInstrumentationTestCase2;
-
-public class FirstTest extends ActivityInstrumentationTestCase2<PlanActivityView> {
-
-    public FirstTest() {
-        super(PlanActivityView.class);
-    }
-
-    public void testExample() {
-        assertEquals(true, true);
-    }
-}
